Fix ICC build break
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@9931 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/IntelFrameworkModulePkg/Universal/BdsDxe/DeviceMngr/DeviceManager.c
@@ -298,7 +298,7 @@ ExtractDisplayedHiiFormFromHiiHandle (
             //
             // Find FormSet OpCode
             //
-            ClassGuidNum = ((EFI_IFR_FORM_SET *) OpCodeData)->Flags & 0x3;
+            ClassGuidNum = (UINT8) (((EFI_IFR_FORM_SET *) OpCodeData)->Flags & 0x3);
             ClassGuid = (EFI_GUID *) (VOID *)(OpCodeData + sizeof (EFI_IFR_FORM_SET));
             while (ClassGuidNum-- > 0) {
               if (CompareGuid (SetupClassGuid, ClassGuid)) {
